Change loading order of configuration backends.
First, try to load an XML configuration file. If that fails, try to create
an XML configuration file and load the plugin store. After the core has
been created (and if the XML could not be loaded), the newly created XML
file is set as configuration on the core, migrating the configuration from
the plugin store to the XML file.